+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * fix(conf): fix error when include inside file (#52)
 * fix(docs): switch M5Stack doc links to english by @lboue (#50)
+* fix(board): fix m5stack coreS3 power issue (#54)
 
 ## v0.1.3 - 2024-06-14